separate in versionX instead of current+legacy
This means that when adding a new version of codecs, we will add a new
file instead of overwriting previous files. It's easier to keep track
of modifications.

Also, the `stateDataCodec` is now more clearly separated.

```
wire
 |
 `-- internal
 |       |
 |       `-- channel
 |       |      |
 |       |      `-- legacy
 |       |      |      |
 |       |      |      `-- version0
 |       |      |      |      |
 |       |      |      |      `-- ChannelCodecs0
 |       |      |      |
 |       |      |      `-- version1
 |       |      |      |      |
 |       |      |      |      `-- ChannelCodecs1
 |       |      |      |
 |       |      |      `-- ChannelTypes
 |       |      |
 |       |      `-- version2
 |       |      |      |
 |       |      |      `-- ChannelCodecs2
 |       |      |
 |       |      `-- ChannelCodecs
 |       |
 |       `-- CommandCodecs
 |
 `-- others
```
