+++ eclair-core/src/main/scala/fr/acinq/eclair/wire/internal/channel/legacy/version0/ChannelCodecs0.scala
@@ -1,4 +1,4 @@
-package fr.acinq.eclair.wire.internal.legacy.legacy0
+package fr.acinq.eclair.wire.internal.channel.legacy.version0
 
 import fr.acinq.bitcoin.DeterministicWallet.{ExtendedPrivateKey, KeyPath}
 import fr.acinq.bitcoin.{ByteVector32, ByteVector64, Crypto, OutPoint, Transaction, TxOut}
@@ -10,7 +10,7 @@ import fr.acinq.eclair.transactions.{CommitmentSpec, DirectedHtlc, IncomingHtlc,
 import fr.acinq.eclair.wire.CommonCodecs.{bytes32, discriminatorWithDefault, publicKey, _}
 import fr.acinq.eclair.wire.LightningMessageCodecs._
 import fr.acinq.eclair.wire.UpdateMessage
-import fr.acinq.eclair.wire.internal.legacy.LegacyChannelTypes
+import fr.acinq.eclair.wire.internal.channel.legacy.ChannelTypes
 import grizzled.slf4j.Logging
 import scodec.bits.BitVector
 import scodec.codecs.{bits, bool, byte, bytes, discriminated, either, ignore, int64, listOfN, optional, peek, provide, uint16, uint32, variableSizeBytes, _}
@@ -24,9 +24,9 @@ import scala.concurrent.duration._
  *
  * Created by PM on 02/06/2017.
  */
-private[internal] object LegacyChannelCodecs0 extends Logging {
+private[channel] object ChannelCodecs0 extends Logging {
 
-  private[legacy0] object Codecs {
+  private[version0] object Codecs {
 
     val keyPathCodec: Codec[KeyPath] = ("path" | listOfN(uint16, uint32)).xmap[KeyPath](l => new KeyPath(l), keyPath => keyPath.path.toList).as[KeyPath].decodeOnly
 
@@ -95,7 +95,7 @@ private[internal] object LegacyChannelCodecs0 extends Logging {
     val txCodec: Codec[Transaction] = variableSizeBytes(uint16, bytes.xmap(d => Transaction.read(d.toArray), d => Transaction.write(d)))
 
     val closingTxCodec: Codec[ClosingTx] = txCodec.decodeOnly.xmap(
-      tx => LegacyChannelTypes.migrateClosingTx(tx),
+      tx => ChannelTypes.migrateClosingTx(tx),
       closingTx => closingTx.tx
     )
 
@@ -232,14 +232,14 @@ private[internal] object LegacyChannelCodecs0 extends Logging {
         ("htlcSuccessTxs" | listOfN(uint16, txCodec)) ::
         ("htlcTimeoutTxs" | listOfN(uint16, txCodec)) ::
         ("claimHtlcDelayedTx" | listOfN(uint16, txCodec)) ::
-        ("spent" | spentMapCodec)).as[LegacyChannelTypes.LegacyLocalCommitPublished].decodeOnly.map[LocalCommitPublished](_.migrate()).decodeOnly
+        ("spent" | spentMapCodec)).as[ChannelTypes.LegacyLocalCommitPublished].decodeOnly.map[LocalCommitPublished](_.migrate()).decodeOnly
 
     val remoteCommitPublishedCodec: Codec[RemoteCommitPublished] = (
       ("commitTx" | txCodec) ::
         ("claimMainOutputTx" | optional(bool, txCodec)) ::
         ("claimHtlcSuccessTxs" | listOfN(uint16, txCodec)) ::
         ("claimHtlcTimeoutTxs" | listOfN(uint16, txCodec)) ::
-        ("spent" | spentMapCodec)).as[LegacyChannelTypes.LegacyRemoteCommitPublished].decodeOnly.map[RemoteCommitPublished](_.migrate()).decodeOnly
+        ("spent" | spentMapCodec)).as[ChannelTypes.LegacyRemoteCommitPublished].decodeOnly.map[RemoteCommitPublished](_.migrate()).decodeOnly
 
     val revokedCommitPublishedCodec: Codec[RevokedCommitPublished] = (
       ("commitTx" | txCodec) ::
@@ -247,7 +247,7 @@ private[internal] object LegacyChannelCodecs0 extends Logging {
         ("mainPenaltyTx" | optional(bool, txCodec)) ::
         ("htlcPenaltyTxs" | listOfN(uint16, txCodec)) ::
         ("claimHtlcDelayedPenaltyTxs" | listOfN(uint16, txCodec)) ::
-        ("spent" | spentMapCodec)).as[LegacyChannelTypes.LegacyRevokedCommitPublished].decodeOnly.map[RevokedCommitPublished](_.migrate()).decodeOnly
+        ("spent" | spentMapCodec)).as[ChannelTypes.LegacyRevokedCommitPublished].decodeOnly.map[RevokedCommitPublished](_.migrate()).decodeOnly
 
     // All channel_announcement's written prior to supporting unknown trailing fields had the same fixed size, because
     // those are the announcements that *we* created and we always used an empty features field, which was the only
