+++ eclair-core/src/test/scala/fr/acinq/eclair/io/SwitchboardSpec.scala
@@ -8,7 +8,7 @@ import fr.acinq.eclair.TestConstants._
 import fr.acinq.eclair.blockchain.TestWallet
 import fr.acinq.eclair.channel.ChannelIdAssigned
 import fr.acinq.eclair.wire._
-import fr.acinq.eclair.wire.internal.ChannelCodecsSpec
+import fr.acinq.eclair.wire.internal.channel.ChannelCodecsSpec
 import fr.acinq.eclair.{Features, NodeParams, TestKitBaseClass, randomBytes32, randomKey}
 import org.scalatest.funsuite.AnyFunSuiteLike
 import scodec.bits._
